Telex_test_app/requirements.txt
Django==4.2.6  
djangorestframework==3.14.0  
django-silk==5.0.3  
pytest==8.0.0
pytest-django==4.8.0
radon==6.0.1  
pylint-django==2.5.5 
gunicorn==21.2.0 
whitenoise==6.6.0 
python-dotenv==1.0.1 
flake8

